Refactor QuestService & prevent non-active user quests progress
This commit introduces two main changes in the QuestService class. Firstly, the creation of quests for newly joined members has been moved into the main class. This centralizes the logic and makes the code more maintainable. Secondly, we've added a check to ensure that if a member is not online, progress isn't added to certain quests. This fixes the issue where quest count would falsely increment for non-active users. The deprecated '_getChannels' function was also removed as part of cleaning up.